debug

+++ semi_supervised.py
@@ -101,9 +101,9 @@ class RandomPointsDataset(Dataset):
         point_indices = [i.to('cpu') for i in point_indices]
         # print(point_indices[0].dtype, point_indices[0].device)
         
-        point_indices_lf = [(F.interpolate(i.unsqueeze(0).unsqueeze(0).to(torch.float64), scale_factor=0.25)).squeeze(0).squeeze(0) for i in point_indices]
-        point_indices = [i.to(torch.int64).to(self.device) for i in point_indices]
-        point_indices_lf = [i.to(torch.int64).to(self.device) for i in point_indices_lf]
+        point_indices_lf = [(F.interpolate(i.unsqueeze(0).unsqueeze(0).to(torch.float32), scale_factor=0.25)).squeeze(0).squeeze(0) for i in point_indices]
+        point_indices = [i.to(torch.int32).to(self.device) for i in point_indices]
+        point_indices_lf = [i.to(torch.int32).to(self.device) for i in point_indices_lf]
 
         # print(point_indices[0].device, point_indices_lf[0].shape)
         # Retrieve image values from selected indices
